Add Yason as a dependency
The proper way for a program to interface with task is `task export`,
this will give me json. YASON is a CL json parser, so I am adding it as
a dependency to process the json.

Switching to `task export` also gives me control over how I display the
task info, which will hopefully make it more meaningful than the current
display format.