Adding projects to *project-list*
Projects are added to *project-list* directly.

The two primary comparison methods, slug-equals and project-equals, are
also created.

Slug-equals checks to see if a slug is equal to a given string, this is
the primary search method for projects.

project-equals checks to see if two projects have the same uuid.
Projects with the same UUID are assumed to be identical.