Correct bug in `add` project function.
Before, it would only work if `area:` and `slug:` were the leading terms
in the string. Now, it works regardless of their position. This is
because we split on whitespace and consider each token in turn. This
also corrected a bug where only the first case of a tag and the first
case of an inherited tag would be added to the list.