Add user config support.
Currently, there are only two configurations worth mention: data-folder
and enable-taskwarrior-integration.

data-folder indicates the location where the projects data is stored.

enable-taskwarrior-integration indicates whether or not the taskwarrior
integration features should be used or not. Currently, this does nothing
as I have not yet wired up the tests for that variable, but that will be
coming soon.

This config file is a lisp file, so it should be edited accordingly.
Instructions can be found at the top of the file.