Remove cl-ascii-text dependency.
cl-ascii-text builds acceptable tables, but they're too bulky for my
tastes. Each table had 3 or 4 lines of separators or deadspace in them.
I have replaced them with a function that builds the table out of format
recipes. This has the added benefit of allowing me to introduce
alternating colors to the table rows, making them easier to read.

Finally, we can customize the reports that come with it as well by
customizing the column labels and functions run by each column.